backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
chromadb
google-generativeai
langchain
langchain-text-splitters
langchain-google-genai
pypdf
python-dotenv
